fix(media_transfer_history): 将 ID 字段类型修改为 uint64 以和数据库 ID 类型匹配

+++ internal/database/media_transfer_history.go
@@ -41,7 +41,7 @@ func QueryMediaTransferHistoryBySrc(src storage.StoragePath) (*models.MediaTrans
 // 如果 ID 为 nil，则根据其他条件查询
 func QueryMediaTransferHistory(
 	ctx context.Context,
-	id *uint, startTime *time.Time, endTime *time.Time,
+	id *uint64, startTime *time.Time, endTime *time.Time,
 	storageType storage.StorageType, // 存储类型和路径，存储类型为 StorageUnknown 时不进行过滤，否则对 src 和 dst 都进行过滤
 	path string, // 路径，模糊匹配
 	transferType storage.TransferType, // 转移类型为 TransferUnknown 时不进行过滤
